% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimation.pqrBayes}
\alias{estimation.pqrBayes}
\title{Estimation and estimation accuracy for a pqrBayes object}
\usage{
estimation.pqrBayes(posterior,coefficient,u.grid,kn,degree,iterations)
}
\arguments{
\item{posterior}{posterior samples from the MCMC.}

\item{coefficient}{the matrix of true varying coefficients evaluated on the grid points.}

\item{u.grid}{the vector of grid points.}

\item{kn}{the number of interior knots for B-spline.}

\item{degree}{the degree of B-spline basis.}

\item{iterations}{the number of MCMC iterations.}
}
\value{
an object of class `pqrBayes.est' is returned, which is a list with components:
\item{error}{integrated mean square errors and total integrated mean square error.}
\item{coeff.est}{estimated values of the varying coefficients.}
}
\description{
calculate estimation and estimation accuracy for varying coefficients
}
\seealso{
\code{\link{pqrBayes}}
}
