% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pqrBayes.R
\name{print.pqrBayes.pred}
\alias{print.pqrBayes.pred}
\title{print a pqrBayes.pred object}
\usage{
\method{print}{pqrBayes.pred}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{pqrBayes.pred object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\value{
No return value, called for side effects.
}
\description{
Print a summary of a pqrBayes.pred object
}
\seealso{
\code{\link{predict_pqrBayes}}
}
