% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pqrBayes.R
\name{print.pqrBayes}
\alias{print.pqrBayes}
\title{print a pqrBayes result}
\usage{
\method{print}{pqrBayes}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{pqrBayes result.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Print a pqrBayes result
}
\seealso{
\code{\link{pqrBayes}}
}
