% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pqrBayes.R
\name{print.pqrBayes.select}
\alias{print.pqrBayes.select}
\title{print a select.pqrBayes object}
\usage{
\method{print}{pqrBayes.select}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{pqrBayes.select object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\value{
No return value, called for side effects.
}
\description{
Print a summary of a select.pqrBayes object
}
\seealso{
\code{\link{pqrBayes.select}}
}
