% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pqrBayes.R
\name{print.VCselect}
\alias{print.VCselect}
\title{print a select.VC object}
\usage{
\method{print}{VCselect}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{VCselect object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\value{
No return value, called for side effects.
}
\description{
Print a summary of a select.VC object
}
\seealso{
\code{\link{VCselect}}
}
