% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Empirical 95\% coverage probability for a pqrBayes object}
\usage{
coverage(posterior,coefficient,u.grid,kn,degree,iterations)
}
\arguments{
\item{posterior}{posterior samples from the MCMC.}

\item{coefficient}{the matrix of true varying coefficients evaluated on the grid points.}

\item{u.grid}{the vector of grid points.}

\item{kn}{the number of interior knots for B-spline.}

\item{degree}{the degree of B-spline basis.}

\item{iterations}{the number of MCMC iterations.}
}
\value{
c
}
\description{
calculate empirical 95\% coverage probabilities for varying coefficients
}
\seealso{
\code{\link{pqrBayes}}
}
