% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqrBayes.R
\name{pqrBayes}
\alias{pqrBayes}
\title{fit Bayesian penalized quantile regression for linear and varying coefficient models}
\usage{
pqrBayes(
  g,
  y,
  u = NULL,
  e,
  quant = 0.5,
  iterations = 10000,
  kn = NULL,
  degree = NULL,
  robust = TRUE,
  sparse = TRUE,
  model = "linear",
  hyper = NULL,
  debugging = FALSE
)
}
\arguments{
\item{g}{the matrix of predictors (subject to selection) without intercept.}

\item{y}{the response variable. The current version only supports the continuous response.}

\item{u}{a vector of effect modifying variable of the quantile varying coefficient model. When fitting a linear model, u = NULL.}

\item{e}{a matrix of clinical covariates (including the intercept if specified) not subject to selection.}

\item{quant}{the quantile level specified by users. The default value is 0.5.}

\item{iterations}{the number of MCMC iterations. The default value is 10,000.}

\item{kn}{the number of interior knots for B-spline. When fitting a linear model, kn = NULL.}

\item{degree}{the degree of B-spline basis. When fitting a linear model, degree = NULL.}

\item{robust}{logical flag. If TRUE, robust methods will be used. Otherwise, non-robust methods will be used. The default value is TRUE.}

\item{sparse}{logical flag. If TRUE, spike-and-slab priors will be adopted to impose exact sparsity on regression coefficients. Otherwise, Laplacian shrinkage will be adopted. The default value is TRUE.}

\item{model}{the model to be fitted. Users can specify "linear" for a linear model and "VC" for a varying coefficient model.}

\item{hyper}{a named list of hyperparameters. The default value is NULL.}

\item{debugging}{logical flag. If TRUE, progress will be output to the console and extra information will be returned. The default value is FALSE.}
}
\value{
an object of class "pqrBayes" is returned, which is a list with components:
\item{obj}{a list of posterior samples from the MCMC and other parameters}
\item{coefficients}{a list of posterior estimates of coefficients}
}
\description{
fit Bayesian penalized quantile regression for linear and varying coefficient models
}
\details{
The quantile regression model described in "\code{\link{data}}" is:
\deqn{Y_{i}=\sum_{k=1}^{q} E_{ik} \beta_k +\sum_{j=0}^{p}X_{ij}\gamma_j +\epsilon_{i},}
where \eqn{\beta_k}'s are the regression coefficients for the clinical covariates (including the intercept if specified) and \eqn{\gamma_j}'s are the regression coefficients of \eqn{\boldsymbol X}.

The quantile varying coefficient model described in "\code{\link{data}}" is:
\deqn{Y_{i}=\sum_{k=1}^{q} E_{ik} \beta_k +\sum_{j=0}^{p}\gamma_j(V_i)X_{ij} +\epsilon_{i},}
where \eqn{\beta_k}'s are the regression coefficients for the clinical covariates and \eqn{\gamma_j}'s are the varying coefficients for the intercept and predictors (e.g. genetic factors).


Users can modify the hyper-parameters by providing a named list of hyper-parameters via the argument `hyper'.
The list can have the following named components
\describe{
  \item{a0, b0: }{ shape parameters of the Beta priors (\eqn{\pi^{a_{0}-1}(1-\pi)^{b_{0}-1}}) on \eqn{\pi_{0}}.}
  \item{c1, c2: }{ the shape parameter and the rate parameter of the Gamma prior on \eqn{\nu}.}
}

Please check the references for more details about the prior distributions.
}
\examples{
## The quantile regression model
data(data)
data = data$data_linear
g=data$g
y=data$y
e=data$e

fit1=pqrBayes(g,y,u=NULL,e,quant=0.5,kn=NULL,degree=NULL,model="linear")
\donttest{

## non-sparse
sparse=FALSE
fit2=pqrBayes(g,y,u=NULL,e,quant=0.5,kn=NULL,degree=NULL,sparse = sparse,model="linear")

## non-robust
robust = FALSE
fit3=pqrBayes(g,y,u=NULL,e,quant=0.5,kn=NULL,degree=NULL,robust = robust,model="linear")

}

## The quantile varying coefficient model
data(data)
data = data$data_varying
g=data$g
y=data$y
u=data$u
e=data$e

fit1=pqrBayes(g,y,u,e,kn=2,degree=2,quant=0.5,model="VC")

\donttest{

## non-sparse
sparse=FALSE
fit2=pqrBayes(g,y,u,e,kn=2,degree=2,quant=0.5,sparse = sparse,model="VC")

## non-robust
robust = FALSE
fit3=pqrBayes(g,y,u,e,kn=2,degree=2,quant=0.5,robust = robust,model="VC")

}
}
\references{
Fan, K., Subedi, S., Yang, G., Lu, X., Ren, J. and Wu, C. (2024). Is Seeing Believing? A Practitioner's Perspective on High-dimensional Statistical Inference in Cancer Genomics Studies. {\emph{Entropy}, 26(9).794} \doi{10.3390/e26090794}

Zhou, F., Ren, J., Ma, S. and Wu, C. (2023). The Bayesian regularized quantile varying coefficient model.
 {\emph{Computational Statistics & Data Analysis}, 107808} \doi{10.1016/j.csda.2023.107808}
 
Ren, J., Zhou, F., Li, X., Ma, S., Jiang, Y. and Wu, C. (2023). Robust Bayesian variable selection for gene-environment interactions. 
{\emph{Biometrics}, 79(2), 684-694} \doi{10.1111/biom.13670}

Ren, J., Zhou, F., Li, X., Chen, Q., Zhang, H., Ma, S., Jiang, Y. and Wu, C. (2020) Semi-parametric Bayesian variable selection for gene-environment interactions.
{\emph{Statistics in Medicine}, 39: 617– 638} \doi{10.1002/sim.8434}
}
\keyword{models}
